refactor: rename to remove sync- prefix
