+++ integrations/xero/tests/xero-bank-transactions.test.ts
@@ -1,15 +1,22 @@
 import { afterEach, vi, expect, it, describe } from 'vitest';
 
-import createSync from '../syncs/sync-accounts.js';
+import createSync from '../syncs/bank-transactions.js';
 
-describe('xero sync-accounts tests', () => {
-    const models = 'Account'.split(',');
+describe('xero bank-transactions tests', () => {
+    const models = 'BankTransaction'.split(',');
 
     const createTestContext = () => {
         const nangoMock = new global.vitest.NangoSyncMock({
             dirname: __dirname,
-            name: 'sync-accounts',
-            Model: 'Account'
+            name: 'bank-transactions',
+            Model: 'BankTransaction'
+        });
+
+        nangoMock.getConnection = vi.fn().mockResolvedValue({
+            connection_config: {
+                tenant_id: '59712f8f-45a3-4d45-a705-5d0c9748317e'
+            },
+            metadata: {}
         });
 
         return {
@@ -25,13 +32,6 @@ describe('xero sync-accounts tests', () => {
 
     it('should get, map correctly the data and batchSave the result', async () => {
         const { nangoMock, batchSaveSpy } = createTestContext();
-
-        nangoMock.getConnection = vi.fn().mockResolvedValue({
-            connection_config: {
-                tenant_id: '59712f8f-45a3-4d45-a705-5d0c9748317e'
-            },
-            metadata: {}
-        });
 
         await createSync.exec(nangoMock);
 
@@ -58,13 +58,6 @@ describe('xero sync-accounts tests', () => {
     it('should get, map correctly the data and batchDelete the result', async () => {
         const { nangoMock } = createTestContext();
         const batchDeleteSpy = vi.spyOn(nangoMock, 'batchDelete');
-
-        nangoMock.getConnection = vi.fn().mockResolvedValue({
-            connection_config: {
-                tenant_id: '59712f8f-45a3-4d45-a705-5d0c9748317e'
-            },
-            metadata: {}
-        });
 
         await createSync.exec(nangoMock);
