+++ integrations/xero/tests/xero-general-ledger.test.ts
@@ -1,15 +1,22 @@
 import { afterEach, vi, expect, it, describe } from 'vitest';
 
-import createSync from '../syncs/sync-items.js';
+import createSync from '../syncs/general-ledger.js';
 
-describe('xero sync-items tests', () => {
-    const models = 'Item'.split(',');
+describe('xero general-ledger tests', () => {
+    const models = 'GeneralLedger'.split(',');
 
     const createTestContext = () => {
         const nangoMock = new global.vitest.NangoSyncMock({
             dirname: __dirname,
-            name: 'sync-items',
-            Model: 'Item'
+            name: 'general-ledger',
+            Model: 'GeneralLedger'
+        });
+
+        nangoMock.getConnection = vi.fn().mockResolvedValue({
+            connection_config: {
+                tenant_id: 'aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee'
+            },
+            metadata: {}
         });
 
         return {
@@ -45,33 +52,6 @@ describe('xero sync-items tests', () => {
             const spied = JSON.parse(JSON.stringify(spiedData));
 
             expect(spied).toStrictEqual(expectedBatchSaveData);
-        }
-    });
-
-    it('should get, map correctly the data and batchDelete the result', async () => {
-        const { nangoMock } = createTestContext();
-        const batchDeleteSpy = vi.spyOn(nangoMock, 'batchDelete');
-
-        await createSync.exec(nangoMock);
-
-        for (const model of models) {
-            const batchDeleteData = await nangoMock.getBatchDeleteData(model);
-            if (batchDeleteData && batchDeleteData.length > 0) {
-                const spiedData = batchDeleteSpy.mock.calls.flatMap((call) => {
-                    if (call[1] === model) {
-                        return call[0];
-                    }
-
-                    return [];
-                });
-
-                // Normalize spy-captured args into plain JSON so they compare cleanly
-                // with fixture data loaded from `*.test.json`.
-                // Removes things like prototypes, undefined values and other non-serializable data.
-                const spied = JSON.parse(JSON.stringify(spiedData));
-
-                expect(spied).toStrictEqual(batchDeleteData);
-            }
         }
     });
 });
